requirements.txt
flask==3.0.3
Werkzeug==3.0.4
flask_cors==5.0.0
google-cloud-storage==2.18.2
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
google-cloud-pubsub